+++ contrib/completions/zoxide.bash
@@ -88,7 +88,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__add)
-            opts="-s -h -V --score --help --version <PATHS>..."
+            opts="-s -h -V --score --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -110,7 +110,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__dedupe)
-            opts="-i -n -h -V --assume-insensitive --dry-run --help --version <pathglob>..."
+            opts="-i -n -h -V --assume-insensitive --dry-run --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -138,7 +138,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__edit__subcmd__decrement)
-            opts="-h -V --help --version <PATH>"
+            opts="-h -V --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -152,7 +152,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__edit__subcmd__delete)
-            opts="-h -V --help --version <PATH>"
+            opts="-h -V --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -166,7 +166,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__edit__subcmd__increment)
-            opts="-h -V --help --version <PATH>"
+            opts="-h -V --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -292,12 +292,16 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__init)
-            opts="-h -V --no-cmd --cmd --hook --help --version bash elvish fish nushell posix powershell tcsh xonsh zsh"
+            opts="-h -V --bind-fzf-insert --no-cmd --cmd --hook --help --version bash elvish fish nushell posix powershell tcsh xonsh zsh"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
             fi
             case "${prev}" in
+                --bind-fzf-insert)
+                    COMPREPLY=($(compgen -f "${cur}"))
+                    return 0
+                    ;;
                 --cmd)
                     COMPREPLY=($(compgen -f "${cur}"))
                     return 0
@@ -314,7 +318,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__query)
-            opts="-a -i -l -s -h -V --all --interactive --list --score --exclude --base-dir --help --version [KEYWORDS]..."
+            opts="-a -i -l -s -h -V --all --interactive --list --score --exclude --base-dir --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0
@@ -342,7 +346,7 @@ _zoxide() {
             return 0
             ;;
         zoxide__subcmd__remove)
-            opts="-h -V --help --version [PATHS]..."
+            opts="-h -V --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0